[test] API 리팩토링에 따른 Test 코드 수정
